Move `LocalFeedImageDataLoader` and `FeedImageDataSource` to production
