+++ EssentialFeedTests/Feed Cache/LocalFeedImageDataLoaderTests.swift
@@ -8,60 +8,6 @@
 import XCTest
 import EssentialFeed
 
-protocol FeedImageDataStore {
-    typealias Result = Swift.Result<Data?, Error>
-
-    func retrieve(dataForURL: URL, completion: @escaping (Result) -> Void)
-}
-
-final class LocalFeedImageDataLoader {
-    private final class Task: FeedImageDataLoaderTask {
-        private var completion: ((FeedImageDataLoader.Result) -> Void)?
-
-        init(completion: (@escaping (FeedImageDataLoader.Result) -> Void)) {
-            self.completion = completion
-        }
-
-        func complete(with result: FeedImageDataLoader.Result) {
-            completion?(result)
-        }
-
-        func cancel() {
-            preventFurtherCompletion()
-        }
-
-        private func preventFurtherCompletion() {
-            completion = nil
-        }
-    }
-
-    public enum Error: Swift.Error {
-        case failed
-        case noFound
-    }
-
-    private let store: FeedImageDataStore
-
-    init(store: FeedImageDataStore) {
-        self.store = store
-    }
-
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void)
-    -> FeedImageDataLoaderTask {
-        let task = Task(completion: completion)
-        store.retrieve(dataForURL: url) { [weak self] result in
-            guard self != nil else { return }
-            
-            task.complete(with: result
-                .mapError { _ in Error.failed }
-                .flatMap { data in
-                    data.map { .success($0) } ?? .failure(Error.noFound)
-                })
-        }
-        return task
-    }
-}
-
 final class LocalFeedImageDataLoaderTests: XCTestCase {
 
     func test_init_doesNotMessageStoreUponCreation() {
